implement aws client interface
