--- pkg/cloud/aws/networks/service.go
@@ -1,64 +1,0 @@
-/*
-Copyright 2024 The Forge contributors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package networks
-
-import (
-	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/forge-build/forge-provider-aws/pkg/cloud"
-)
-
-type vpcsInterface interface {
-	CreateVpc(input *ec2.CreateVpcInput) (*ec2.CreateVpcOutput, error)
-	DeleteVpc(input *ec2.DeleteVpcInput) (*ec2.DeleteVpcOutput, error)
-	DescribeVpcs(input *ec2.DescribeVpcsInput) (*ec2.DescribeVpcsOutput, error)
-	DescribeRouteTables(input *ec2.DescribeRouteTablesInput) (*ec2.DescribeRouteTablesOutput, error)
-	CreateRoute(input *ec2.CreateRouteInput) (*ec2.CreateRouteOutput, error)
-}
-
-type internetGatewaysInterface interface {
-	DetachInternetGateway(input *ec2.DetachInternetGatewayInput) (*ec2.DetachInternetGatewayOutput, error)
-	AttachInternetGateway(input *ec2.AttachInternetGatewayInput) (*ec2.AttachInternetGatewayOutput, error)
-	CreateInternetGateway(input *ec2.CreateInternetGatewayInput) (*ec2.CreateInternetGatewayOutput, error)
-	DeleteInternetGateway(input *ec2.DeleteInternetGatewayInput) (*ec2.DeleteInternetGatewayOutput, error)
-	DescribeInternetGateways(input *ec2.DescribeInternetGatewaysInput) (*ec2.DescribeInternetGatewaysOutput, error)
-}
-
-type Scope interface {
-	cloud.Build
-	Region() string
-	VPCSpec() *ec2.CreateVpcInput
-	VPCID() *string
-	VPCName() *string
-}
-
-// Service implements networks reconciler.
-type Service struct {
-	scope     Scope
-	ec2Client vpcsInterface
-	igwClient internetGatewaysInterface
-}
-
-var _ cloud.Reconciler = &Service{}
-
-// New returns Service from given scope.
-func New(scope Scope) *Service {
-	return &Service{
-		scope:     scope,
-		ec2Client: scope.Cloud(),
-		igwClient: scope.Cloud(),
-	}
-}
