+++ pkg/controllers/awsbuild/controller.go
@@ -25,13 +25,13 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/forge-build/forge-provider-aws/pkg/cloud"
-	awserrors "github.com/forge-build/forge-provider-aws/pkg/cloud/aws/errors"
-	"github.com/forge-build/forge-provider-aws/pkg/cloud/aws/images"
-	"github.com/forge-build/forge-provider-aws/pkg/cloud/aws/instances"
-	"github.com/forge-build/forge-provider-aws/pkg/cloud/aws/networks"
-	"github.com/forge-build/forge-provider-aws/pkg/cloud/aws/securitygroup"
-	"github.com/forge-build/forge-provider-aws/pkg/cloud/aws/subnet"
 	"github.com/forge-build/forge-provider-aws/pkg/cloud/scope"
+	awserrors "github.com/forge-build/forge-provider-aws/pkg/cloud/services/errors"
+	"github.com/forge-build/forge-provider-aws/pkg/cloud/services/images"
+	"github.com/forge-build/forge-provider-aws/pkg/cloud/services/instances"
+	"github.com/forge-build/forge-provider-aws/pkg/cloud/services/networks"
+	"github.com/forge-build/forge-provider-aws/pkg/cloud/services/securitygroup"
+	"github.com/forge-build/forge-provider-aws/pkg/cloud/services/subnet"
 	buildv1 "github.com/forge-build/forge/api/v1alpha1"
 	"github.com/forge-build/forge/pkg/ssh"
 	forgeutil "github.com/forge-build/forge/util"
